Refactor prompt registry microservice  (#1124)
