+++ .github/workflows/docker/compose/prompt_registry-compose.yaml
@@ -3,7 +3,7 @@
 
 # this file should be run in the root of the repo
 services:
-  promptregistry-mongo-server:
+  promptregistry-server:
     build:
-      dockerfile: comps/prompt_registry/mongo/Dockerfile
-    image: ${REGISTRY:-opea}/promptregistry-mongo-server:${TAG:-latest}
+      dockerfile: comps/prompt_registry/src/Dockerfile
+    image: ${REGISTRY:-opea}/promptregistry-server:${TAG:-latest}
