+++ comps/prompt_registry/src/integrations/mongo/__init__.py
@@ -1,0 +1,2 @@
+# Copyright (C) 2024 Intel Corporation
+# SPDX-License-Identifier: Apache-2.0
