+++ comps/prompt_registry/src/integrations/mongo/config.py
