+++ comps/prompt_registry/src/prompt_store.py
@@ -4,8 +4,8 @@
 
 import bson.errors as BsonError
 from bson.objectid import ObjectId
-from config import COLLECTION_NAME
-from mongo_conn import MongoClient
+from integrations.mongo.config import COLLECTION_NAME
+from integrations.mongo.mongo_conn import MongoClient
 
 
 class PromptStore:
@@ -16,8 +16,9 @@ class PromptStore:
     ):
         self.user = user
 
-    def initialize_storage(self) -> None:
-        self.db_client = MongoClient.get_db_client()
+    def initialize_storage(self, db_type="mongo") -> None:
+        if db_type == "mongo":
+            self.db_client = MongoClient.get_db_client()
         self.collection = self.db_client[COLLECTION_NAME]
 
     async def save_prompt(self, prompt) -> str:
