+++ comps/prompt_registry/src/requirements.txt
